admin function added
admin functionality

+++ views/film-details.tpl
@@ -13,8 +13,10 @@
 			<div class="card__header">
 				<h4 class="title-4"> <?=$film['name'];?> </h4>
 				<div class="buttons">
-					<a href="edit-film.php?id=<?=$film['id'];?>" class="button button--edit">Изменить</a>
-					<a href="index.php?action=delete&id=<?=$film['id'];?>" class="button button--remove">Удалить</a>
+					<?php if ( isUserExist() ) { ?>
+						<a href="edit-film.php?id=<?=$film['id'];?>" class="button button--edit">Изменить</a>
+						<a href="index.php?action=delete&id=<?=$film['id'];?>" class="button button--remove">Удалить</a>
+					<?php } ?>
 				</div>
 			</div>
 			<div class="badge"> <?php echo $film['genre'];?> </div>
